[fix] wrong imports after refactoring
